requirements.txt
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
xlrd==1.2.0
